Rename CommentAction to OvertimeCommentAction
