+++ src/main/java/org/synyx/urlaubsverwaltung/overtime/OvertimeCommentAction.java
@@ -5,7 +5,7 @@ package org.synyx.urlaubsverwaltung.overtime;
  *
  * @since 2.11.0
  */
-public enum OvertimeAction {
+public enum OvertimeCommentAction {
 
     CREATED,
     EDITED
